+++ packages/components/src/components/tag-dismissible/tag-dismissible-utils.ts
@@ -1,12 +1,4 @@
-/** @deprecated */
-export const TAG_DISMISSIBLE_COLORS_DEPRECATED = ['background-default'] as const; // 'background-default' is deprecated (replaced with 'background-base')
-/** @deprecated */
-export type TagDismissibleColorDeprecated = (typeof TAG_DISMISSIBLE_COLORS_DEPRECATED)[number];
-export const TAG_DISMISSIBLE_COLORS = [
-  'background-base',
-  'background-surface',
-  ...TAG_DISMISSIBLE_COLORS_DEPRECATED,
-] as const;
+export const TAG_DISMISSIBLE_COLORS = ['background-base', 'background-surface'] as const;
 export type TagDismissibleColor = (typeof TAG_DISMISSIBLE_COLORS)[number];
 
 export const TAG_DISMISSIBLE_ARIA_ATTRIBUTES = ['aria-label'] as const;