+++ packages/components/src/components/tag/tag-shared-utils.ts
@@ -1,11 +1,8 @@
 import type { ThemedColors } from '../../styles';
-import type { TagColor, TagColorDeprecated } from './tag-utils';
+import type { TagColor } from './tag-utils';
 
-export const getThemedBackgroundColor = (
-  tagColor: Exclude<TagColor, TagColorDeprecated>,
-  themedColors: ThemedColors
-): string => {
-  const colorMap: Record<Exclude<TagColor, TagColorDeprecated>, string> = {
+export const getThemedBackgroundColor = (tagColor: TagColor, themedColors: ThemedColors): string => {
+  const colorMap: Record<TagColor, string> = {
     'background-base': themedColors.backgroundColor,
     'background-surface': themedColors.backgroundSurfaceColor,
     'background-frosted': themedColors.backgroundFrostedColor,