+++ packages/components/src/components/tag/tag-utils.ts
@@ -1,20 +1,10 @@
-import type { IconName, Theme } from '../../types';
 import type { ThemedColors } from '../../styles'; // deep import needed since barrel contains MutationObserver and causes VRT to fail because of TAG_COLORS import
-import { TAG_DISMISSIBLE_COLORS, type TagDismissibleColorDeprecated } from '../tag-dismissible/tag-dismissible-utils';
+import type { IconName, Theme } from '../../types';
 import { darken, isThemeDark, lighten } from '../../utils';
+import { TAG_DISMISSIBLE_COLORS } from '../tag-dismissible/tag-dismissible-utils';
 
 export type TagIcon = IconName;
 
-/** @deprecated */
-export const TAG_COLORS_DEPRECATED = [
-  'neutral-contrast-high', // 'notification-contrast-high' is deprecated (replaced with 'primary')
-  'notification-neutral', // 'notification-neutral' is deprecated (replaced with 'notification-info-soft')
-  'notification-warning',
-  'notification-success',
-  'notification-error',
-] as const;
-/** @deprecated */
-export type TagColorDeprecated = (typeof TAG_COLORS_DEPRECATED)[number] | TagDismissibleColorDeprecated;
 export const TAG_COLORS = [
   ...TAG_DISMISSIBLE_COLORS,
   'background-frosted',
@@ -23,18 +13,13 @@ export const TAG_COLORS = [
   'notification-warning-soft',
   'notification-success-soft',
   'notification-error-soft',
-  ...TAG_COLORS_DEPRECATED,
 ] as const;
 export type TagColor = (typeof TAG_COLORS)[number];
 
-export const getThemedBackgroundHoverColor = (
-  tagColor: Exclude<TagColor, TagColorDeprecated>,
-  themedColors: ThemedColors,
-  theme: Theme
-): string => {
+export const getThemedBackgroundHoverColor = (tagColor: TagColor, themedColors: ThemedColors, theme: Theme): string => {
   const isDark = isThemeDark(theme);
   const keySuffix = isDark ? 'Lighten' : 'Darken';
-  const colorMap: Record<Exclude<TagColor, TagColorDeprecated>, string> = {
+  const colorMap: Record<TagColor, string> = {
     'background-base': themedColors[`backgroundColor${keySuffix}`],
     'background-surface': themedColors[`backgroundSurfaceColor${keySuffix}`],
     'background-frosted': isDark